+++ HISTORY.rst
@@ -15,6 +15,7 @@ Unreleased
 **Features and Improvements**
 
 * ALCYN-1495: Add the entry register report just after the delivery report (only if required)
+* ALCYN-1544: Improve response time of customer 2 years stat web service (ESB).
 
 **Bugfixes**